Bump React to 16.3.2 to fix IE11 warning
React 16.3.2 uses the `Map( iterable )` constructor
to create a map of valid Fragment props. This is a
problem in IE11 because IE11 supports a form of Map
but not the constructor taking the iterable. The result
is that the map is empty and React warns that `children`
is not a valid Fragment prop. React 16.3.1 and later
contain a fix for this, and I'm bumping to React 16.3.2
as the latest 16.3 version.